feat: Refactor ClaudeNode component and implement useClaudeNode composable for improved state management and UI components
